requirements.txt
pydantic==2.10.1
httpx==0.27.2
llama-index-core
llama-index-llms-openai
llama-index-llms-replicate
llama-index-embeddings-huggingface